Small change to main, allowed for full details to be viewed on account
Small change to main, allowed for full details to be viewed on account. Made it so that it shows account details whether you're a tutor or student. Ignore big blocks of green and red, legit was just copy, cut , paste, etc. All I changed on main was adding a logout function to tutor and deleted the full account details